Introduce the 'macvtap_mode' nicparam
A MacVTap device can operate in one of four modes which are defined at
creation time and determine how the tap endpoints communicate between
each other. Those modes are the 'vepa', 'bridge', 'private', and the
'passthru' mode.

This patch creates a new nicparam slot named 'macvtap_mode', which will
specify the operation mode of a MacVTap NIC. The macvtap_mode parameter
will only be used if the network's connection mode is set to macvtap,
since it does not make sense in other modes. Finally we've created a
separate constant variable for each one of the four macvtap modes.