Create two new helpers in 'lib/query.py' module
This patch introduces two new helper methods to the 'lib/query.py'
module, to be used for queries regarding to MacVTap NICs.

The first method is the '_GetInstNicMacvtapMode' and returns the
'macvtap_mode' nicparam of a given MacVTap NIC, while the second one,
named '_GetInstAllNicMacvtapModes', returns the 'macvtap_mode' from all
the MacVTap NICs of an instance.