Introduce the 'ComputeMacvtapModeNicParam' method
This new helper method will be used to validate and compute the
macvtap_mode nicparam of a NIC using the MacVTap device driver.

It accepts a single argument, the macvtap_mode nicparam, and checks
whether it is a valid MacVTap mode, i.e., vepa, bridge, private,
passthru, or not. If it isn't, an OpPrereqError exception will be
raised. Finally, in case of an empty macvtap_mode argument, it will be
set to the default MacVTap operation mode for Ganeti, the bridge mode.