Rename 'bridges_exist' RPC call to 'netdevs_exist'
This patch, modifies the 'bridges_exist' RPC to support MacVTap devices
too.

Currently, the 'bridges_exist' RPC call is used to check whether the
given bridge exists on the given node. This check is necessary for the
lower device of the MacVTap interface too, and so we extend that RPC
call to also include this interface. The check is done using the sysfs.

Since the updated RPC call will support more than one device types we
rename it to 'netdevs_exist' and also we update the respective methods'
documentation and invocation wherever it is required.