Generate TAP names for the macvtap interfaces
This patch extends the 'GenerateTapName' method of the hv_base module in
order to generate TAP names for macvtap interfaces too. These names are
of the form 'gnt.macvtap.%d' where '%d' is a unique number within the
node.

To do so, we've added an extra boolean argument to the 'GenerateTapName'
method which specifies whether this method is used for NICs in instance
communication mode (the default case), or not.