Introduce the OpenMacVTap method for KVMhypervisor
This method will create and open a macvtap device using the `ip-link`
command.

It takes as arguments the name of the device to be created, its mac
address, the macvtap mode, the link where the macvtap device will be
attached to, and a dictionary of features to be enabled for the device.
Currently, these features are vhost, and multiqueue. The vnet_hdr is
enabled by default and there is no way to disable it, after creating the
device.

Also, we've created a helper method to get the fds of the /dev/vhost-net
device, since that part of code will be used by both the OpenMacVTap and
OpenTap methods.