Introduce the 'ListInstancesMacvtapNICs' method
This method will return a list with the MacVTap NICs (if any) of the
running KVM instances. Currently, it is implemented for the KVM
hypervisor only since MacVTap initial functionality targets the KVM
hypervisor. For the rest hypervisors a HypervisorError exception will be
raised if this method is called.

This is an essential method in order to be able to clean up stale
MacVTap devices caused by a user instance shutdown, as stated by the
next patch.