watcher: Add a method to cleanup MacVTap devices
MacVTap devices are persistent. In some cases, like an instance internal
user shutdown with the kvm daemon disabled, the instance's MacVTap NICs
will not be cleaned up. So, we've created a new method inside the
ganeti-watcher, named '_CleanupStaleMacvtapDevs', which will remove all
stale MacVTap NICs that are not used by an instance.

To find those NICs, we compute the difference of all the Ganeti macvtap
devices on the node the watcher is running ("gnt.macvtap." prefix) and
the macvtap devs that are in use by an instance (based on the runtime
nic files).