Add an extra arg to 'HotModDevice' function
Currently, hot-modifying a device requires to firstly hot-delete the
device and then re-add it using the updated NIC parameters. However, the
device object passed to the HotDel method is the one with the updated
NIC params instead of the previous params, resulting in passing wrong
environment values to the configuration NIC scripts.

This patch fixes that behavior by adding an extra argument to the
HotModDevice function, specifying the old device object.